% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableTechniques.R
\name{availableTechinques}
\alias{availableTechinques}
\title{Available techniques}
\usage{
availableTechinques()
}
\value{
A character vector of techniques.
}
\description{
List techniques available for processing.
}
