% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableTechniques.R
\name{availableTechniques}
\alias{availableTechniques}
\title{Available techniques}
\usage{
availableTechniques()
}
\value{
A character vector of techniques.
}
\description{
List techniques available for processing.
}
\examples{
availableTechniques()
}
