% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{extractProcObject}
\alias{extractProcObject}
\alias{extractProcObject,MetaboProfile-method}
\title{extractProcObject}
\usage{
extractProcObject(x)

\S4method{extractProcObject}{MetaboProfile}(x)
}
\arguments{
\item{x}{S4 object of class MetaboProfile}
}
\description{
Extract processing package object from an object of class MetaboProfile
}
