% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\docType{methods}
\name{extractXCMSnExp}
\alias{extractXCMSnExp}
\alias{extractXCMSnExp,MetaboProfile-method}
\title{extractXCMSnExp}
\usage{
extractXCMSnExp(x)

\S4method{extractXCMSnExp}{MetaboProfile}(x)
}
\arguments{
\item{x}{S4 object of class MetaboProfile}
}
\description{
Extract XCMSnExp objects from an object of class MetaboProfile
}
