% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{MetaboProfile-class}
\alias{MetaboProfile-class}
\title{MetaboProfile}
\description{
An S4 class to store the profile processing results
}
\section{Slots}{

\describe{
\item{\code{version}}{package version}

\item{\code{creation_date}}{date and time of the initiation of processing}

\item{\code{files}}{list of file paths to used for processing. Vectors of files form different aquisition modes should be labelled accordingly.}

\item{\code{processing_parameters}}{object of class ProfileParameters containing the parameters for processing}

\item{\code{Info}}{tibble containing runinfo data}

\item{\code{Data}}{list containing tibbles of processed data}

\item{\code{processingResults}}{list containing processing results}
}}

