% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{technique}
\alias{technique}
\alias{technique<-}
\alias{processingParameters}
\alias{processingParameters<-}
\alias{technique,ProfileParameters-method}
\alias{technique<-,ProfileParameters-method}
\alias{processingParameters,ProfileParameters-method}
\alias{processingParameters<-,ProfileParameters-method}
\title{ProfileParameters getters and setters}
\usage{
technique(x)

technique(x) <- value

processingParameters(x)

processingParameters(x) <- value

\S4method{technique}{ProfileParameters}(x)

\S4method{technique}{ProfileParameters}(x) <- value

\S4method{processingParameters}{ProfileParameters}(x)

\S4method{processingParameters}{ProfileParameters}(x) <- value
}
\arguments{
\item{x}{S4 object of class ProfileParameters}

\item{value}{value to set}
}
\description{
Get and set profile processing parameters
}
