% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\docType{methods}
\name{peakInfo}
\alias{peakInfo}
\alias{peakInfo,MetaboProfile-method}
\title{peakInfo}
\usage{
peakInfo(x)

\S4method{peakInfo}{MetaboProfile}(x)
}
\arguments{
\item{x}{S4 object of class MetaboProfile}
}
\description{
Extract peak information from an object of class MetaboProfile
}
