% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotting.R
\docType{methods}
\name{plotChromatogram}
\alias{plotChromatogram}
\alias{plotChromatogram,MetaboProfile-method}
\title{plotChromatogram}
\usage{
plotChromatogram(processed, cls = NULL, group = F, alpha = 1,
  aggregationFun = "max", ...)

\S4method{plotChromatogram}{MetaboProfile}(processed, cls = NULL,
  group = F, alpha = 1, aggregationFun = "max", ...)
}
\arguments{
\item{processed}{S4 object of class MetaboProfile}

\item{cls}{sample information column to use for plot colours.}

\item{group}{average samples within groups. If cls is NULL, plot average chromatogram across all samples.}

\item{alpha}{plot line transparancy alpha}

\item{aggregationFun}{value to pass to \code{xcms::chromatogram()} for chromatogram type. Defaults to base peak ("max").}

\item{...}{arguments to pass to \code{erah::plotChr()}}
}
\description{
Plot ion chromatogram from MetaboProfile.
}
