% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotting.R
\name{plotTIC}
\alias{plotTIC}
\alias{plotTIC,MetaboProfile-method}
\title{plotTIC}
\usage{
plotTIC(processed, by = "injOrder", colour = "block")

\S4method{plotTIC}{MetaboProfile}(processed, by = "injOrder", colour = "block")
}
\arguments{
\item{processed}{S4 object of class MetaboProfile}

\item{by}{info column to plot against}

\item{colour}{info column to provide colour labels}
}
\description{
Plot sample total ion counts.
}
