% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{processedData}
\alias{processedData}
\alias{processedData,MetaboProfile-method}
\title{processedData}
\usage{
processedData(x)

\S4method{processedData}{MetaboProfile}(x)
}
\arguments{
\item{x}{S4 object of class MetaboProfile}
}
\description{
Extract processed data from an object of class MetaboProfile
}
