% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{processingParameters}
\alias{processingParameters}
\alias{processingParameters<-}
\alias{processingParameters,MetaboProfile-method}
\alias{processingParameters<-,MetaboProfile-method}
\title{Get or set processing parameters}
\usage{
processingParameters(x)

processingParameters(x) <- value

\S4method{processingParameters}{MetaboProfile}(x)

\S4method{processingParameters}{MetaboProfile}(x) <- value
}
\arguments{
\item{x}{S4 object of class MetaboProfile}

\item{value}{list object of parameters to set}
}
\value{
A list containing processing parameter objects.
}
\description{
Get or set processing parameters for an S4 object of class MetaboProfile.
}
