% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ProfileParameters-class}
\alias{ProfileParameters-class}
\title{ProfileParameters}
\description{
An S4 class to store profile processing parameters.
}
\section{Slots}{

\describe{
\item{\code{technique}}{the profiling technique to use. Defaults to the result of \code{availableTechniques()[1]}}

\item{\code{processing_parameters}}{a list containing the parameters to use for processing}
}}

