% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileParameters.R
\name{profileParameters}
\alias{profileParameters}
\title{Profiling paramters}
\usage{
profileParameters(technique)
}
\arguments{
\item{technique}{the profiling technique for which to initiate parameters}
}
\value{
An S4 object of class ProfileParameters.
}
\description{
Initiate default processing parameters for a given profiling technique.
}
\examples{
profileParameters('LCMS-RP')
}
\seealso{
\code{\link{availableTechniques}}, \code{\link{ProfileParameters-class}}, \code{\link{profileProcess}}
}
