% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileProcess.R
\name{profileProcess}
\alias{profileProcess}
\title{profileProcess}
\usage{
profileProcess(files, info, parameters)
}
\arguments{
\item{files}{character vector of file paths to use for processing}

\item{info}{tibble containing sample info}

\item{parameters}{object of class ProfileParameters containing the parameters for processing}
}
\description{
process metabolomic profiling data
}
\examples{
\dontrun{
# Get sample information and use only the three injection for this example
info <- metaboData::runinfo('RP-UHPLC-HRMS','BdistachyonEcotypes') \%>\%
  dplyr::filter(injOrder \%in\% seq_len(3))

# Retrieve file paths
files <- metaboData::filePaths('RP-UHPLC-HRMS','BdistachyonEcotypes') \%>\%
  {
    .[base::basename(.) \%in\% info$fileName] 
  }

# Prepare parameters
parameters <- profileParameters('LCMS-RP')

# Run processing
pd <- profileProcess(files,info,parameters)
}
}
