% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{sampleInfo}
\alias{sampleInfo}
\alias{sampleInfo,MetaboProfile-method}
\title{sampleInfo}
\usage{
sampleInfo(x)

\S4method{sampleInfo}{MetaboProfile}(x)
}
\arguments{
\item{x}{S4 object of class MetaboProfile}
}
\description{
Extract sample info from an object of class MetaboProfile.
}
