% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,MetaboProfile-method}
\alias{show,MetaboProfile-method}
\title{show-MetaboProfile}
\usage{
\S4method{show}{MetaboProfile}(object)
}
\arguments{
\item{object}{S4 object of class MetaboProfile}
}
\description{
show method for MetaboProfile class
}
