% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,ProfileParameters-method}
\alias{show,ProfileParameters-method}
\title{show-ProfileParameters}
\usage{
\S4method{show}{ProfileParameters}(object)
}
\arguments{
\item{object}{S4 object of class ProfileParameters}
}
\description{
show method for ProfileParameters class
}
