% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{sampleInfo}
\alias{sampleInfo}
\alias{sampleInfo<-}
\alias{processedData}
\alias{processedData<-}
\alias{extractProcObject}
\alias{peakInfo}
\alias{sampleInfo,MetaboProfile-method}
\alias{sampleInfo<-,MetaboProfile-method}
\alias{processedData,MetaboProfile-method}
\alias{processedData<-,MetaboProfile-method}
\alias{extractProcObject,MetaboProfile-method}
\alias{peakInfo,MetaboProfile-method}
\title{MetaboProfile class get and set methods.}
\usage{
sampleInfo(x)

sampleInfo(x) <- value

processedData(x)

processedData(x) <- value

extractProcObject(x)

peakInfo(x)

\S4method{sampleInfo}{MetaboProfile}(x)

\S4method{sampleInfo}{MetaboProfile}(x) <- value

\S4method{processedData}{MetaboProfile}(x)

\S4method{processedData}{MetaboProfile}(x) <- value

\S4method{extractProcObject}{MetaboProfile}(x)

\S4method{peakInfo}{MetaboProfile}(x)
}
\arguments{
\item{x}{S4 object of class MetaboProfile}

\item{value}{value to set}
}
\description{
Retrieve or set information for a MetaboProfile object.
}
