% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{version}
\alias{version}
\alias{creationDate}
\alias{filePaths}
\alias{filePaths<-}
\alias{sampleInfo}
\alias{processingResults}
\alias{processingResults<-}
\alias{processedData}
\alias{processedData<-}
\alias{extractProcObject}
\alias{peakInfo}
\alias{version,MetaboProfile-method}
\alias{creationDate,MetaboProfile-method}
\alias{filePaths,MetaboProfile-method}
\alias{filePaths<-,MetaboProfile-method}
\alias{sampleInfo,MetaboProfile-method}
\alias{processingResults,MetaboProfile-method}
\alias{processingResults<-,MetaboProfile-method}
\alias{processedData,MetaboProfile-method}
\alias{processedData<-,MetaboProfile-method}
\alias{extractProcObject,MetaboProfile-method}
\alias{peakInfo,MetaboProfile-method}
\title{MetaboProfile class get and set methods.}
\usage{
version(x)

creationDate(x)

filePaths(x)

filePaths(x) <- value

sampleInfo(x)

processingResults(x)

processingResults(x) <- value

processedData(x)

processedData(x) <- value

extractProcObject(x)

peakInfo(x)

\S4method{version}{MetaboProfile}(x)

\S4method{creationDate}{MetaboProfile}(x)

\S4method{filePaths}{MetaboProfile}(x)

\S4method{filePaths}{MetaboProfile}(x) <- value

\S4method{sampleInfo}{MetaboProfile}(x)

\S4method{processingResults}{MetaboProfile}(x)

\S4method{processingResults}{MetaboProfile}(x) <- value

\S4method{processedData}{MetaboProfile}(x)

\S4method{processedData}{MetaboProfile}(x) <- value

\S4method{extractProcObject}{MetaboProfile}(x)

\S4method{peakInfo}{MetaboProfile}(x)
}
\arguments{
\item{x}{S4 object of class MetaboProfile}

\item{value}{value to set}
}
\description{
Retrieve or set information for a MetaboProfile object.
}
