% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ProfileParameters-class}
\alias{ProfileParameters-class}
\title{ProfileParameters}
\description{
An S4 class to store profile processing parameters.
}
\section{Slots}{

\describe{
\item{\code{technique}}{the profiling technique to use}

\item{\code{infoName}}{name of file containing run information}

\item{\code{processingParameters}}{a list containing the parameters to use for processing}
}}

