% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileParameters.R
\name{profileParameters}
\alias{profileParameters}
\title{profileParameters}
\usage{
profileParameters(technique = NULL)
}
\arguments{
\item{technique}{the profiling technique for which to initiate parameters}
}
\description{
Initiate default processing parameters for a given profiling technique.
}
