% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileProcess.R
\name{profileProcess}
\alias{profileProcess}
\title{profileProcess}
\usage{
profileProcess(files, info, parameters)
}
\arguments{
\item{files}{character vector of file paths to use for processing}

\item{info}{file path for info file}

\item{parameters}{object of class ProfileParameters containing the parameters for processing}
}
\description{
process metabolomic profiling data
}
