% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileProcess.R
\name{profileProcess}
\alias{profileProcess}
\title{profileProcess}
\usage{
profileProcess(files, info, parameters)
}
\arguments{
\item{files}{character vector of file paths to use for processing}

\item{info}{file path for info file}

\item{parameters}{object of class ProfileParameters containing the parameters for processing}
}
\description{
process metabolomic profiling data
}
\examples{
\dontrun{
library(stringr)
library(faahKO)
library(tibble)
library(readr)

files <- list.files(system.file("cdf", package = "faahKO"),full.names = TRUE,recursive = T)

# make runinfo file
filesSplit <- str_split_fixed(files,'/',str_count(files[1],'/') + 1) 
info <- tibble(fileOrder = 1:length(files),
              injOrder = 1:length(files),
              fileName = filesSplit[,ncol(filesSplit)], 
              batch = rep(1,length(files)), 
              batchBlock = rep(1,length(files)), 
              names = str_replace_all(filesSplit[,ncol(filesSplit)],'.CDF',''), 
              class = filesSplit[,ncol(filesSplit)-1])
system('mkdir .tmp')
write_csv(info,'.tmp/runinfo.csv')

files <- list(pos = files)
info <- '.tmp/runinfo.csv'

# prepare parameters
parameters <- profileParameters('LCMS-RP')
parameters@processingParameters$peakDetection <- CentWaveParam(snthresh = 20, noise = 1000)
parameters@processingParameters$retentionTimeCorrection <- ObiwarpParam()
parameters@processingParameters$grouping <- PeakDensityParam(maxFeatures = 300,minFraction = 2/3)
# run processing
processedData <- profileProcess(files,info,parameters)

# tidy up
system('rm -r .tmp')
}
}
