% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileProcess.R
\name{profileProcess}
\alias{profileProcess}
\title{profileProcess}
\usage{
profileProcess(files, parameters)
}
\arguments{
\item{files}{character vector of file paths to use for processing along with the info file path}

\item{parameters}{object of class ProfileParameters containing the parameters for processing}
}
\description{
process metabolomic profiling data
}
